Add files via upload
The main page in LearnEase application receives fresh upgrades from the "UI and Usability Enhancements" commit which involves a complete revamp of LearningPage.xaml and LearningPage.xaml.cs.The most important part of these upgrades is introduction of new interface and interaction designs which provide smooth and user-friendly experience. A conversion has been performed on the home page that has changed it from boring appearance to an engaging one by the use of a cutting-edge layout that captures the attention of a student and leads them to learn with keen interest. In this perception the best part of it is that the user experience is better thanks to the increased accessibility and the ease to use the platform as a result, which leads to fewer obstacles the learners have to face. This page features colorful graphics and pictures that tell the story of an excellent app. Improving the user experience by making the app easy to use is also the aim of this page.