Add files via upload
Implement constructor in CourseDetailPage.xaml.cs: Added a new constructor to accept courseName as a parameter. This constructor initializes the page with specific course details, setting up the framework for displaying tailored content to the user. The update paves the way for a more dynamic and personalized user experience, allowing for course-specific information to be rendered on the CourseDetailPage. This change marks a significant improvement in how course details are handled and displayed within the application.