Add files via upload
Introduce Constructor in CourseDetailPage.xaml.cs: Integrate a novel constructor to accommodate the courseName parameter. This constructor serves as the gateway to tailor the page with specific course particulars, setting the stage for a personalized user experience. By enabling the dynamic rendering of course-specific information, this update heralds a new era in how course details are presented within the application.